Auth & protected routes

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 // Controllers folder holds CRUD functions:
 const asyncHandler = require("express-async-handler"); // Await response from DB.
 const Goal = require("../models/goalModel"); // goalSchema for MongoDB
+const User = require("../models/userModel"); // userSchema for MongoDB
 
 // @description: Read goals
 // @route: GET /api/goals
 // @access: Private with authentication
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
 });
@@ -15,7 +16,7 @@ const getGoals = asyncHandler(async (req, res) => {
 // @route: POST /api/goal
 // @access: Private with authentication
 const createGoal = asyncHandler(async (req, res) => {
-    console.log(req.body); // Using Postman POST body.text with `urlencoded` middleware:
+    // console.log(req.body); // Using Postman POST body.text with `urlencoded` middleware:
     if (!req.body.text) {
         res.status(400);
         throw new Error("Please add a text field."); // Express error handler.
@@ -23,6 +24,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     });
 
     res.status(200).json(goal);
@@ -37,6 +39,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found.");
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // Check for user:
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found.");
+    }
+
+    // Check logged in user matches the goal user id:
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("Not authenticated.");
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +71,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found.");
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // Check for user:
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found.");
+    }
+
+    // Check logged in user matches the goal user id:
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("Not authenticated.");
     }
 
     await goal.deleteOne();
